Updates and CUDA 10 testing (#197)
* Adding updated notes on Python build, GOOFIT_VERSION

* Move to using Dim5

* Minor notes

* Much stricter requirement on Valid Minuit2 convergence
