+++ src/PDFs/physics/Amp4Body.cu
@@ -33,7 +33,7 @@ class.
 #include <goofit/PDFs/physics/Amp4BodyGlobals.h>
 #include <goofit/PDFs/physics/Amplitude.h>
 #include <goofit/PDFs/physics/detail/AmpCalc.h>
-#include <goofit/PDFs/physics/detail/EvalVar.h>
+#include <goofit/PDFs/physics/detail/Dim5.h>
 #include <goofit/PDFs/physics/detail/LSCalculator.h>
 #include <goofit/PDFs/physics/detail/NormIntegrator.h>
 #include <goofit/PDFs/physics/detail/NormLSCalculator.h>
