+++ src/PDFs/physics/detail/NormIntegrator_TD.cu
@@ -3,7 +3,7 @@
 #include <goofit/PDFs/ParameterContainer.h>
 #include <goofit/PDFs/physics/Amp4BodyGlobals.h>
 #include <goofit/PDFs/physics/SpinFactors.h>
-#include <goofit/PDFs/physics/detail/EvalVar.h>
+#include <goofit/PDFs/physics/detail/Dim5.h>
 #include <goofit/detail/Complex.h>
 
 #include <thrust/functional.h>
